[#10] Allow timeout for messages
fix #10

Messages that should be transmitted through a direct connection are now
stored until the connection to the recipient is open (a request for a
direct connection is emitted when the message is stored).

Stored messages can have a timeout value after which they are removed from
the queue. If a callback function was associated with the message, this
callback is then triggered.